WebSockets/tests/submissions/Opera/opening-handshake/001.html

<!doctype html>
<title>WebSockets: invalid handshake</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
setTestTimeout(10000, 'got no close event');
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':8007/invalid');
  ws.onclose = function(e) {
    assert_equals(e.wasClean, false);
    debug(e);
    ws.onclose = assertUnreached;
    setTestTimeout(null);
    setTimeout(end, 50);
  }
   ws.onmessage = ws.onopen = assertUnreached;
});
</script>
